refactor: 유저 도메인 리팩토링 끝!!
