+++ src/main/java/com/meoguri/linkocean/domain/user/entity/User.java
@@ -15,15 +15,15 @@ import javax.persistence.UniqueConstraint;
 
 import com.meoguri.linkocean.domain.BaseIdEntity;
 import com.meoguri.linkocean.domain.profile.entity.Profile;
+import com.meoguri.linkocean.domain.user.entity.vo.Email;
+import com.meoguri.linkocean.domain.user.entity.vo.OAuthType;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 /**
  * 사용자
- *
- * - 사용자는 단순 로그인 용도로 쓰이고 실질적으로 프로필이 사용자를 대체한다.
- * - 사용자를 등록할 때 [이메일, 소셜 타입]이 존재해야 한다.
+ * - 사용자를 등록할 때 [이메일, 소셜 타입]은 필수이다.
  */
 @Getter
 @NoArgsConstructor(access = PROTECTED)
@@ -45,9 +45,7 @@ public class User extends BaseIdEntity {
 	@OneToOne(fetch = LAZY, mappedBy = "user")
 	private Profile profile;
 
-	/**
-	 * 회원 가입시 사용하는 생성자
-	 */
+	/* 회원 가입시 사용하는 생성자 */
 	public User(final String email, final String oauthType) {
 		checkNotNull(email);
 		checkNotNull(oauthType);
@@ -59,18 +57,4 @@ public class User extends BaseIdEntity {
 	public Long getProfileId() {
 		return profile == null ? null : profile.getId();
 	}
-
-	/**
-	 * OAuth 지원 벤더 종류
-	 * GOOGLE, NAVER, KAKAO
-	 */
-	public enum OAuthType {
-
-		GOOGLE, NAVER, KAKAO;
-
-		public static OAuthType of(final String type) {
-			return OAuthType.valueOf(type);
-		}
-	}
-
 }
