+++ src/main/java/com/meoguri/linkocean/domain/user/service/UserService.java
@@ -3,9 +3,9 @@ package com.meoguri.linkocean.domain.user.service;
 import org.springframework.stereotype.Service;
 
 import com.meoguri.linkocean.configuration.security.jwt.JwtProvider;
-import com.meoguri.linkocean.domain.user.entity.Email;
 import com.meoguri.linkocean.domain.user.entity.User;
-import com.meoguri.linkocean.domain.user.entity.User.OAuthType;
+import com.meoguri.linkocean.domain.user.entity.vo.Email;
+import com.meoguri.linkocean.domain.user.entity.vo.OAuthType;
 import com.meoguri.linkocean.domain.user.repository.UserRepository;
 
 import lombok.RequiredArgsConstructor;
